A new List parameter *resources* has been added to the worker. This parameter lists resources the worker has. These will be checked by the scheduler on work request and based on that tasks will be held back from execution by workers not having the required resources. * Task resource requirements (*worker_resources*) are reported by the worker to   the scheduler on _add_task() * The list of available resources (*worker_resources*) is sent by the worker on   requesting new work on _get_work()
The scheduler stores worker resource requirements *worker_resources* with each
task. On work request *get_work()*, the scheduler checks if the decision is
trivial (this was changed to consider the existence of worker resource
requirements as well):
* only check tasks reported by the worker requesting
* check all tasks reported by all workers

If the scheduler finds a suitable task (even if it is not scheduled by the
worker requesting), it will send it back, as the worker can load the task
dynamically. This implies that the in_workers (a bool value showing thether the
worker has the found task scheduled or not) check was removed from the
scheduler code.